feat: 심화 과제 초기 설정
