+++ src/advanced/components/ui/Button.tsx
@@ -1,0 +1,29 @@
+interface ButtonProps {
+  children?: React.ReactNode;
+  onClick?: () => void;
+  disabled?: boolean;
+  variant?: 'primary' | 'secondary' | 'danger';
+  className?: string;
+  icon?: React.ReactNode;
+  type?: 'button' | 'submit' | 'reset';
+}
+
+export const Button = ({ children, onClick, disabled, variant = 'primary', className = '', icon }: ButtonProps) => {
+  const baseClasses = 'px-4 py-2 rounded-md font-medium transition-colors flex items-center justify-center';
+  const variantClasses = {
+    primary: 'bg-gray-900 text-white hover:bg-gray-800',
+    secondary: 'bg-gray-100 text-gray-900 hover:bg-gray-200',
+    danger: 'bg-red-600 text-white hover:bg-red-700'
+  };
+
+  return (
+    <button
+      onClick={onClick}
+      disabled={disabled}
+      className={`${baseClasses} ${variantClasses[variant]} ${className}`}
+    >
+      {icon && <span className="mr-2">{icon}</span>}
+      {children}
+    </button>
+  );
+};
